backend/app-builder/code/chapter12/requirements.txt
Flask-AppBuilder==4.3.0
pytest==7.2.0
pytest-cov==4.0.0
pytest-flask==1.2.0
pytest-html==3.2.0